Initial implementation of store wrapper to allow resource req pattern to be used
Big issue is that use of store requires significant rewrite of most resource-based code, and this can lead to errors creeping in around resource release.

This adds a new store class - the VidigiStore - that can use the familiar .request() and with: syntax, hopefully minimizing the work required to incorporate vidigi into existing simpy projects.