Make column names flexible in all functions (addresses #36) as part of work for #51
Column names were hardcoded to 'minutes', which made the logic hard to follow when adding in more flexibility to support a wider range of time unit displays. Have made the naming of time columns more unit-agnostic, and also allowed flexibility in naming of the core columns required in the input dataframes.